moved helper functions into helper file
