+++ automationHelpers.js
@@ -1,0 +1,108 @@
+const knex = require("knex");
+const { DATABASE_URL } = require('./src/config');
+
+//  pg returns numeric values as strings
+//  this converts all numeric types to floats (decimal)
+var types = require('pg').types
+types.setTypeParser(1700, function(val) {
+  return parseFloat(val)
+})
+
+const db = knex({
+  client: 'pg',
+  connection: DATABASE_URL,
+});
+
+const selectGoals = async params => {
+  return await db('goals')
+    .select(
+      'goals.id',
+      'goals.name',
+      'user_id',
+      'goal_amount',
+      'contribution_amount',
+      'current_amount',
+      'end_date',
+      'completed',
+      'users.allowance',
+      'users.balance',
+    )
+    .where(params)
+    .join('users', {'goals.user_id': 'users.id'});
+}
+
+const selectUserAllowance = async id => {
+  const result =  await db('users')
+    .select('allowance')
+    .where({ id })
+    .first()
+
+    return result.allowance
+}
+
+const selectGoal = async id => {
+  return await db('goals')
+    .select()
+    .where({ id })
+    .first()
+}
+
+const updateGoal = async (id, params) => {
+  return await db('goals')
+    .where({ id })
+    .update(params)
+}
+
+const createAlert = async (user_id, complete, name) => {
+  return await db('alerts')
+    .insert({
+      'user_id': user_id,
+      'title': complete ? 'Goal Complete!' : 'Insufficient Allowance.',
+      'message': complete ? `You completed your goal, ${name}` : `Looks like you don't have enough allowance to fund your goal, ${name}`,
+    })
+}
+
+const selectUserAlerts = async (user_id) => {
+  const alerts = await db('alerts')
+    .select()
+    .where({ user_id })
+  
+    console.log(alerts)
+}
+
+const moveContribution = async (goal, allowance, adjusted) => {
+  
+  const difference = goal.goal_amount - goal.contribution_amount;
+  
+  // subtract contribution amount from allowance
+  adjusted
+    ? allowance -= difference // if true, subtract difference from allowance 
+    : allowance -= goal.contribution_amount  //if false, subtract contribution amount from allowance
+
+  // add contribution amount to goal's current amount
+  adjusted
+    ? goal.current_amount += difference //if true, add difference to current amt
+    : goal.current_amount += goal.contribution_amount // if false, add contribution amt to current amt
+  
+  await db.transaction(async trx => {
+    // update allowance value on users table
+    await trx('users')
+      .where({ 'id': goal.user_id})
+      .update({ allowance: allowance })
+    
+    // update current amount in goals table
+    await updateGoal(goal.user_id, { 'current_amount': goal.current_amount })
+  });
+  
+  const g = await selectGoal(goal.id) //dev
+  console.log(g.name, g.current_amount, g.completed)
+}
+
+module.exports = {
+  selectGoals,
+  updateGoal,
+  selectUserAlerts,
+  createAlert,
+  selectUserAllowance,
+  moveContribution,
+}
