--- scheduled-goals-service.js
@@ -1,5 +1,0 @@
-const scheduledGoalsService = {
-  getGoals(db) {
-    return db('goals').where('completed', '=', 'false')
-  }
-}
